+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Replace `<img>` tags in posts with links.
-- Generate Ed25519 keys for new accounts.
+- Generate Ed25519 keys for all accounts.
 
 ### Changed
 
@@ -18,7 +18,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
-- Prevent `add-ed25591-key` command from overwriting existing key.
 - Don't return error when replying to a public post with a direct message.
 
 ## [2.8.0] - 2024-01-29